requirements.txt
bs4
requests
datetime
python-decouple
flask
gunicorn